Improved handling of sporadic mouse inputs.
An example of mouse input that would break the handler is
pressing left mouse button in the terminal and releasing it
outside the terminal. In this case, the terminal thinks no button
mouse move events are now left mouse button move events.
In fact, the terminal will still think that, but before methods
on `_PRESSED_KEYS` would error and handling could break.
Now, to fix the mouse-button state, user just needs to re-click the left
mouse button. Handling should be more robust to these errors.