`"reverse"` field added to Cell.
So before this change, if one wanted, say, an upper 5/8ths block (which doesn't exist), a lower 3/8ths block (`"▃"`) could instead be used with the foreground and background colors swapped.

The issue with this is alpha compositing always composites with the background color! So the flipped foreground and background
colors will look sorely out-of-place as soon as any transparency is
enabled.

`"reverse"` fixes this. Now, an upper 5/8ths block can be created
by setting the cell's `"reverse"` field and alpha compositing will
still look correct using the `"bg_color"` field.