vt100_terminal.py: fix (?) mouse support on Linux
This patch restores batgrl's ability to work with:

- xterm
- konsole
- kitty
- ghostty

I wish I could explain why changing the order of these commands is
necessary! This patch was authored through trial and error. However, it
just so happens to bring back enable_mouse_support to the contents it
had in e726ea8de003fb3f0fd9c59dc2bc2c8929b051ea. It appears that this
might be a bug that was introduced in the refactoring of commit
2f594edae4388f1502b78789aa31957abd264dfe ("v0.35.0") and made
enable_mouse_support and disable_mouse_support stop looking alike.